感染状況・医療提供体制の分析: Rename components/MonitoringCommentCard.vue to components/index/SiteTopUpper/MonitoringComment/Card.vue
