+++ components/index/SiteTopUpper.vue
@@ -15,8 +15,8 @@ import Vue from 'vue'
 
 import Consultation from '@/components/Consultation.vue'
 import InfectionMedicalCareProvisionStatus from '@/components/index/SiteTopUpper/InfectionMedicalCareProvisionStatus.vue'
+import MonitoringCommentCard from '@/components/index/SiteTopUpper/MonitoringComment/Card.vue'
 import WhatsNew from '@/components/index/SiteTopUpper/WhatsNew.vue'
-import MonitoringCommentCard from '@/components/MonitoringCommentCard.vue'
 import StayingPopulation from '@/components/StayingPopulation.vue'
 import {
   News as INews,
